v1.7 - Added Engineer class and changes to Tester

+++ aircraft/Operations.java
@@ -2,21 +2,21 @@ package aircraft;
 
 public class Operations {
     // Inspect/ operation number
-    private float id;
+    private double id;
     // Frequency broken up into hours and date
-    private float frequencyH;
-    private float frequencyD;
+    private double frequencyH;
+    private double frequencyD;
     // Hours already logged
-    private float hours;
+    private double hours;
     // The due date for the component
     private String date;
     // Value of the tachometer
-    private float tachometer;
+    private double tachometer;
     // Any note left by the engineer
     private String note;
 
     // Constructor method for the case where both frequencies are specified
-    public Operations(int id, float frequencyH, float frequencyD, float hours, String date, String note){
+    public Operations(int id, double frequencyH, double frequencyD, double hours, String date, String note){
         this.id = id;
         this.frequencyH = frequencyH;
         this.frequencyD = frequencyD;
@@ -26,7 +26,7 @@ public class Operations {
     }
 
     // Constructor method for the case where only one type of frequency is specified
-    public Operations(int id, float frequency, float hours, String date, String note){
+    public Operations(int id, double frequency, double hours, String date, String note){
         this.id = id;
         if (frequency < 50) {
             this.frequencyD = frequency;
@@ -42,22 +42,22 @@ public class Operations {
     // GETTER METHODS FOR ALL VARIABLES
 
     // Method to get the ID
-    public float getID(){
+    public double getID(){
         return this.id;
     }
 
     // Method to get the frequency for the days
-    public float getFrequencyD(){
+    public double getFrequencyD(){
         return this.frequencyD;
     }
 
     // Method to get the frequency for the hours
-    public float getFrequencyH(){
+    public double getFrequencyH(){
         return this.frequencyH;
     }
 
     // Method to get hours
-    public float getHours(){
+    public double getHours(){
         return this.hours;
     }
 
@@ -67,7 +67,7 @@ public class Operations {
     }
 
     // Method to get tachometer value
-    public float getTachometer(){
+    public double getTachometer(){
         return this.tachometer;
     }
 
@@ -82,15 +82,15 @@ public class Operations {
     }
 
     // Method to update frequency
-    public void updateFrequency(float frequencyH, float frequencyD){
+    public void updateFrequency(double frequencyH, double frequencyD){
         this.frequencyH = frequencyH;
         this.frequencyD = frequencyD;
     }
 
     // Method to calculate the hours remaining before a part needs to change
-    public float hoursRemaining(float tachometer){
-        float tachDifference = tachometer - this.tachometer;
-        float hoursRemaining = this.hours - tachDifference;
+    public double hoursRemaining(double tachometer){
+        double tachDifference = tachometer - this.tachometer;
+        double hoursRemaining = this.hours - tachDifference;
         /* If tachDifference is greater than hours then print statement needs to be made to notify the user that the part urgently needs
         to be changed */
         return hoursRemaining;
